Typos (and small comment)
Pattern matching:
"The syntax of using pattern matching is similar to defining ordinary functions, but instead of using variable names, you use the values."
  and
"Note that you can pattern match on a variable too! Variable is like a pattern that matches any value and gives it a name. You can think of variables in function definitions as special cases of pattern matching."
==> Formulation is a little confusing.