Small typos (and comment)
I like the quote you put in from Jasper Van der Jeugt. It was the typo in his name that prompted me to fork. : )
I had read through this tutorial and it was clear that many of the 'typos' had their origin in the fact that your native language is Russian. But I found that quite charming. : ) I did think that some of them might be a bit 'jarring' for some readers but so be it. However, it is never bad to make the experience 'smoother'.
Earlier I saw the keynote by Simon Peyton Jones at Scala eXchange 2020 and SPJ /also/ wrote his name wrong (different error though). Attributions are not difficult to get right.
But it must be something with Dutch/Flemish that the rest of the world finds difficult. I have stopped using my own name in international contexts for over 2 decades. You can call me whatever you like, but if you use my actual name I do appreciate it be written correctly. (Or pronounced correctly ... which my American wife still can't do ... so I changed my name to help her out. : )
Feel free to ignore/reject all my edits, but please keep the correction on Jasper's name. : )